ai_engine/requirements.txt
google-generativeai>=0.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
